More cleanup and internal doc
Removed skipping mapped files with no packages.  This can occur
during updating, but there's no reasn not to re-calculate all checksums
for all mapped files.  If we add any additional functionality in the future
which would require those checksums to be updated to match file-state
mid-update, it could actually be desired.