Made checksums map appropriately.
This ensures that if you keep a file you're always tracking the new
checksum.  If the new checksum is equal to the one in the opus map
it will never prompt you to resolve, because you've already answered
that question in the past.

Only when the file changes will it prompt you.  This will cut down on
the number of conflicts you need to resolve (i.e. instead of one for
every file, it should only ever be the ones which have changed).

This is the appropriate behavior.